requirements.txt
pyinotify==0.9.6
tweepy==3.5.0
python-dotenv==0.6.1